Improve integration tests
